Revert "[Play-190] Radio Kit Typescript Conversion (#1928)"
This reverts commit 4140fa0c3ed39786b57f7413101bef51a8315247.
